+++ browser/html/typeTagSvgStop.go
@@ -2410,7 +2410,7 @@ func (e *TagSvgStop) Offset(value interface{}) (ref *TagSvgStop) {
 	}
 
 	if converted, ok := value.(float64); ok {
-		p := strconv.FormatFloat(float64(converted), 'g', -1, 64)
+		p := strconv.FormatFloat(converted, 'g', -1, 64)
 		e.selfElement.Call("setAttribute", "offset", p)
 		return e
 	}
